+++ server/src/common/actions/effectifs.actions.ts
@@ -11,6 +11,7 @@ import { defaultValuesEffectif } from "@/common/model/effectifs.model/effectifs.
 import { stripEmptyFields } from "../utils/miscUtils";
 
 import { legacySchema } from "./effectif.legacy_schema";
+import { createComputedStatutObject } from "./effectifs.statut.actions";
 
 /**
  * Méthode de build d'un effectif
@@ -122,19 +123,35 @@ export const lockEffectif = async (effectif: IEffectif) => {
   return updated.value;
 };
 
-export const addEffectifComputedFields = (organisme: IOrganisme): IEffectif["_computed"] => {
-  return {
-    organisme: {
-      ...(organisme.adresse?.region ? { region: organisme.adresse.region } : {}),
-      ...(organisme.adresse?.departement ? { departement: organisme.adresse.departement } : {}),
-      ...(organisme.adresse?.academie ? { academie: organisme.adresse.academie } : {}),
-      ...(organisme.adresse?.bassinEmploi ? { bassinEmploi: organisme.adresse.bassinEmploi } : {}),
-      ...(organisme.uai ? { uai: organisme.uai } : {}),
-      ...(organisme.siret ? { siret: organisme.siret } : {}),
-      ...(organisme.reseaux ? { reseaux: organisme.reseaux } : {}),
-      fiable: organisme.fiabilisation_statut === "FIABLE" && !organisme.ferme,
-    },
-  };
+export const addComputedFields = ({
+  organisme,
+  effectif,
+}: {
+  organisme?: IOrganisme;
+  effectif?: IEffectif;
+}): Partial<IEffectif["_computed"]> => {
+  const computedFields: Partial<IEffectif["_computed"]> = {};
+
+  if (organisme) {
+    const { adresse, uai, siret, reseaux, fiabilisation_statut, ferme } = organisme;
+    computedFields.organisme = {
+      ...(adresse?.region && { region: adresse?.region }),
+      ...(adresse?.departement && { departement: adresse?.departement }),
+      ...(adresse?.academie && { academie: adresse?.academie }),
+      ...(adresse?.bassinEmploi && { bassinEmploi: adresse?.bassinEmploi }),
+      ...(uai && { uai }),
+      ...(siret && { siret }),
+      ...(reseaux && { reseaux }),
+      fiable: fiabilisation_statut === "FIABLE" && !ferme,
+    };
+  }
+
+  if (effectif) {
+    const statut = createComputedStatutObject(effectif, new Date());
+    computedFields.statut = statut;
+  }
+
+  return computedFields;
 };
 
 export async function getEffectifForm(effectifId: ObjectId): Promise<any> {
